Added TSMNet class, added some changes to main
Resolve #7
